pyproject.toml
[project]
name = "pdf-parsing-solution"
version = "0.1.0"
description = "PDF parsing solution: Docling + YOLO + VLM"
readme = "README.md"
requires-python = ">=3.10, <3.13"
authors = [{ name = "pySwag" }]
dependencies = [
    "docling>=2.90.0",
    "paddleocr>=3.5.0",
    "paddlepaddle>=3.3.1",
    "torch>=2.11.0",
    "transformers>=5.5.4",
    "ultralytics>=8.4.40",
]

[project.scripts]
run-parser = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "utils"]
